ベクトル検索部分を別ファイルに切り出し
